+++ Nakama/Source/NakamaCore/Private/NakamaCoreClientFactory.cpp
@@ -1,7 +1,7 @@
 #include "NakamaCoreClientFactory.h"
 #include "UnrealLogSink.h"
 #include "http.h"
-#include "websocket.h"
+#include "UnrealWsTransport.h"
 #include "WebSocketsModule.h"
 
 void NakamaCoreClientFactory::initLogging(Nakama::NLogLevel level)
@@ -19,9 +19,3 @@ Nakama::NRtClientPtr NakamaCoreClientFactory::createNakamaRtClient(const Nakama:
 {
 	return client->createRtClient(Nakama::NRtTransportPtr(new Nakama::Unreal::UnrealWsTransport()));
 }
-
-Nakama::NRtClientPtr NakamaCoreClientFactory::createNakamaRtClient(const Nakama::NClientPtr& client, FWebSocketsModule* websocketsModule)
-{
-	return client->createRtClient(Nakama::NRtTransportPtr(new Nakama::Unreal::UnrealWsTransport(websocketsModule)));
-}
-